Split BaseCaptureVC up into three extension files
