+++ Source/UI/Capture/View Controllers/BaseCaptureViewController.swift
@@ -11,9 +11,7 @@ import Cartography
 import RealmSwift
 
 fileprivate typealias LocalClass = BaseCaptureViewController
-class BaseCaptureViewController: UIViewController,
-DebugBarDelegate, UIGestureRecognizerDelegate, ComponentMenuBarDelegate,
-ComponentEditBarDelegate {
+class BaseCaptureViewController: UIViewController, UIGestureRecognizerDelegate {
     
     lazy var kit: Kit = { KitManager.currentKit }()
     
@@ -60,7 +58,7 @@ ComponentEditBarDelegate {
     
     // MARK: - Capture Button Setup
     
-    private lazy var captureButton: CaptureButton = CaptureButton()
+    lazy var captureButton: CaptureButton = CaptureButton()
     fileprivate func captureButtonSetup() {
         captureButton.setTappedCallback(instance: self,
                                         method: Method.captureTapped)
@@ -78,36 +76,8 @@ ComponentEditBarDelegate {
     
     // MARK: - Component Menu Setup
     
-    fileprivate lazy var menuView = ComponentMenuBar()
-    fileprivate var menuVerticalConstraint: NSLayoutConstraint?
-    fileprivate func controlMenuSetup() {
-        menuView.delegate = self
-        let blurEffect = UIBlurEffect(style: .light)
-        let visualEffectView = UIVisualEffectView(effect: blurEffect)
-        menuView.addSubview(visualEffectView)
-        constrain(visualEffectView) {
-            let superview = $0.superview!
-            $0.edges == superview.edges
-        }
-        
-        menuView.backgroundColor = UIColor.clear
-        view.addSubview(menuView)
-        
-        var verticalConstraint: NSLayoutConstraint!
-        constrain(menuView) {
-            let superview = $0.superview!
-            $0.left == superview.left
-            $0.height == 55
-            $0.width == superview.width
-            $0.top >= superview.top ~ 1000
-            $0.bottom <= superview.bottom ~ 1000
-            verticalConstraint = $0.top == superview.top + 100
-            verticalConstraint ~ 400
-        }
-        
-        menuVerticalConstraint = verticalConstraint
-        menuDraggableSetup()
-    }
+    lazy var menuView = ComponentMenuBar()
+    var menuVerticalConstraint: NSLayoutConstraint?
     
     lazy var dragGesture: UIPanGestureRecognizer = {
         let dragGesture = UIPanGestureRecognizer(target: self,
@@ -115,10 +85,6 @@ ComponentEditBarDelegate {
         dragGesture.delegate = self
         return dragGesture
     }()
-    
-    fileprivate func menuDraggableSetup() {
-        menuView.addGestureRecognizer(dragGesture)
-    }
     
     func gestureRecognizerShouldBegin(_ gestureRecognizer: UIGestureRecognizer) -> Bool {
         if gestureRecognizer == dragGesture {
@@ -140,39 +106,7 @@ ComponentEditBarDelegate {
         }
     }
     
-    fileprivate var menuLastTranslationY: CGFloat?
-    func menuDragged(gesture: UIPanGestureRecognizer) {
-        if gesture.state == .began || gesture.state == .ended || gesture.state == .cancelled {
-            menuLastTranslationY = nil
-        }
-        guard gesture.state == .changed else {
-            return
-        }
-        
-        let translation = gesture.translation(in: menuView)
-        var translationY = translation.y
-        
-        if let menuLastTranslationY = menuLastTranslationY {
-            translationY -= menuLastTranslationY
-        }
-        
-        let minimumOffsetChange: CGFloat = 1
-        guard abs(translationY) > minimumOffsetChange,
-        let menuVerticalConstraint = menuVerticalConstraint else {
-            return
-        }
-        
-        menuLastTranslationY = translation.y
-        
-        menuVerticalConstraint.constant += translationY
-        menuView.setNeedsLayout()
-    }
-    
-    // MARK: - Capturing
-    
-    func captureTapped() {
-        print("capture tapped")
-    }
+    var menuLastTranslationY: CGFloat?
     
     // MARK: - Status Bar
     
@@ -203,61 +137,6 @@ ComponentEditBarDelegate {
         }
     }
     
-    // MARK: - Debug Bar
-    
-    var barItems: [DebugBarItem] {
-        let jitterKey = DebugBarItem(title: "Jitter")
-        jitterKey.tapClosure = {
-            self.captureButton.startKeyTimesJitter()
-        }
-        
-        let stopJitter = DebugBarItem(title: "Stop Jittter")
-        stopJitter.tapClosure = {
-            self.captureButton.stopJittter()
-        }
-        
-        let clearComponents = DebugBarItem(title: "Clear Components")
-        clearComponents.tapClosure = { [unowned self] in
-            self.removeAllComponents()
-        }
-        
-        let switchCamera = DebugBarItem(title: "Switch Camera")
-        switchCamera.tapClosure = { [unowned self] in
-            self.switchCamera()
-        }
-        
-        return [jitterKey, stopJitter, clearComponents, switchCamera]
-    }
-    
-    func removeAllComponents() {
-        for component in kit.components {
-            component.view.removeFromSuperview()
-        }
-        
-        kit.components.removeAll()
-        kit.saveKit()
-    }
-        
-    func switchCamera() {
-        print("Switch camera!")
-    }
-    
-    // MARK: - Component Menu Bar Delegate
-    
-    func add(component: Component.Type,
-             atFrame frame: CGRect, fromView view: UIView) {
-        let componentInstance = component.createInstance()
-        let componentView = componentInstance.view
-        componentInstance.frame = view.convert(frame, to: self.view)
-        
-        self.view.insertSubview(componentView, belowSubview: menuView)
-        
-        kit.addComponent(component: componentInstance)
-        kit.saveKit()
-        
-        configureAdded(component: componentInstance)
-    }
-    
     // MARK: - Load Kit
     
     func loadComponents() {
@@ -266,30 +145,6 @@ ComponentEditBarDelegate {
             self.view.insertSubview(component.view, belowSubview: menuView)
             configureAdded(component: component)
         }
-    }
-    
-    func addEditGesture(toComponent component: Component) {
-        let gesture = UILongPressGestureRecognizer(target: self, action: .componentEditGesture)
-        component.view.addGestureRecognizer(gesture)
-    }
-    
-    func componentEditGesture(gesture: UILongPressGestureRecognizer) {
-        guard let targetView = gesture.view else {
-            fatalError("No view associated with gesture!")
-        }
-        
-        guard gesture.state == .began else {
-            return
-        }
-        
-        for component in kit.components where targetView == component.view {
-            configureEditBar(withTargetComponent: component)
-            // remove hold to edit gesture
-            targetView.removeGestureRecognizer(gesture)
-            return
-        }
-        
-        fatalError("Edit gesture failed")
     }
     
     // MARK: - Component Configuring
@@ -305,152 +160,32 @@ ComponentEditBarDelegate {
     // MARK: - Edit Components
     
     lazy var editBar = ComponentEditBar()
-    fileprivate var editBarVerticalConstraint: NSLayoutConstraint?
-    fileprivate func componentEditBarSetup() {
-        view.addSubview(editBar)
-        var verticalConstraint: NSLayoutConstraint!
-        constrain(editBar) {
-            let superview = $0.superview!
-            $0.left == superview.left
-            $0.height == 100
-            $0.width == superview.width
-            $0.top >= superview.top ~ 1000
-            $0.bottom <= superview.bottom ~ 1000
-            verticalConstraint = $0.top == superview.top + 300
-            verticalConstraint ~ 400
-        }
-        editBar.delegate = self
-        editBar.isHidden = true
-        editBarVerticalConstraint = verticalConstraint
-        editBarDraggableSetup()
-    }
+    var editBarVerticalConstraint: NSLayoutConstraint?
     
-    fileprivate lazy var editBarDragGesture: UIPanGestureRecognizer = {
+    lazy var editBarDragGesture: UIPanGestureRecognizer = {
         let dragGesture = UIPanGestureRecognizer(target: self,
                                                  action: .editBarDragged)
         dragGesture.delegate = self
         return dragGesture
     }()
     
-    fileprivate func editBarDraggableSetup() {
-        editBar.addGestureRecognizer(editBarDragGesture)
-    }
-    
-    fileprivate var editBarLastTranslationY: CGFloat?
-    func editBarDragged(gesture: UIPanGestureRecognizer) {
-        if gesture.state == .began || gesture.state == .ended || gesture.state == .cancelled {
-            editBarLastTranslationY = nil
-        }
-        guard gesture.state == .changed else {
-            return
-        }
-        
-        let translation = gesture.translation(in: editBar)
-        var translationY = translation.y
-        
-        if let editBarLastTranslationY = editBarLastTranslationY {
-            translationY -= editBarLastTranslationY
-        }
-        
-        let minimumOffsetChange: CGFloat = 1
-        guard abs(translationY) > minimumOffsetChange,
-            let editBarVerticalConstraint = editBarVerticalConstraint else {
-                return
-        }
-        
-        editBarLastTranslationY = translation.y
-        
-        editBarVerticalConstraint.constant += translationY
-        editBar.setNeedsLayout()
-    }
-    
-    fileprivate func loadComponentEditBar() {
-        let editControl = CircleSlider()
-        self.view.addSubview(editControl)
-        
-        constrain(editControl) {
-            let superview = $0.superview!
-            $0.center == superview.center
-            $0.width == 50
-            $0.height == $0.width
-        }
-    }
-    
-    fileprivate func configureEditBar(withTargetComponent target: Component) {
-        editBar.isHidden = false
-        editBar.set(target: target)
-        
-        editGestures.removeAll()
-        if target is EditPosition {
-            addPositionEditGesture(toComponent: target)
-        }
-    }
+    var editBarLastTranslationY: CGFloat?
     
     // MARK: - Edit Position
     
     var editGestures = [DragGesture]()
-    func addPositionEditGesture(toComponent targetComponent: Component) {
-        guard let component = targetComponent as? EditPosition else {
-            return
-        }
-        
-        let gesture = DragGesture(withView: targetComponent.view)
-        editGestures.append(gesture)
-        gesture.dragHandler = { difference in
-            var origin = component.origin
-            origin.x += difference.x
-            origin.y += difference.y
-            component.origin = origin
-        }
+    
+    // MARK: - Camera Actions
+    
+    func switchCamera() {
+        print("Switch camera!")
     }
     
-    // MARK: - Component Edit Bar Delegate
-    
-    func cancel(editingComponent: Component) {
-        editGestures.removeAll()
-        addEditGesture(toComponent: editingComponent)
-    }
-    
-    func save(component: Component) {
-        editGestures.removeAll()
-        addEditGesture(toComponent: component)
-        kit.saveKit()
-    }
-    
-    func askUserForDeleteConfirmation(component: Component,
-                                      userConfirmedBlock: @escaping ComponentEditBarDelegate.UserConfirmedDeleteBlock) {
-        let alertController = UIAlertController(title: "Delete Component",
-                                                message: "Are you sure you want to delete this component?",
-                                                preferredStyle: .alert)
-        let cancel = UIAlertAction(title: "Cancel", style: .cancel)
-        alertController.addAction(cancel)
-        
-        let kit = self.kit
-        let delete = UIAlertAction(title: "Delete", style: .destructive) { action in
-            userConfirmedBlock(true)
-            self.editGestures.removeAll()
-            component.view.removeFromSuperview()
-            guard let index = kit.components.index(where: {$0 === component}) else {
-                fatalError("Couldn't find component to delete: \(component)")
-            }
-            kit.components.remove(at: index)
-            kit.saveKit()
-        }
-        alertController.addAction(delete)
-        
-        present(alertController, animated: true, completion: nil)
+    func captureTapped() {
+        print("capture tapped")
     }
 }
 
-extension Realm {
-    func filter<ParentType: Object>(parentType: ParentType.Type,
-                                    subclasses: [ParentType.Type],
-                                    predicate: NSPredicate) -> [ParentType] {
-        return ([parentType] + subclasses).flatMap { classType in
-            return Array(self.objects(classType).filter(predicate))
-        }
-    }
-}
 // MARK: - Callbacks
 
 private struct Method {
